requirements.txt
fastapi 
uvicorn 
crewai 
langchain-google-genai 
pygithub 
python-dotenv